Address PR feedback: macOS MainWindow partial, typed menu click handlers, static file picker lists, split StringResource ternary
